Separate out key frame request from MAX_INT frame ACK
A call to rdpClientConProcessMsgClientRegionEx() from xrdp with the frame
number set to MAX_INT acks all outstanding frames. This is useful when
a decoder has been deleted for a resize.

A second feature of this call is that it ensures the next frame sent is
a key frame in progressive mode.

This is fine for a single monitor system. On a multi-monitor system
however, this logic for this breaks one of the assumptions made by
rdpDeferredUpdateCallback() which is that only updates for a single
monitor are sent at once. In extreme output situations, this can result
in some corruption on a multi-monitor resize.

This PR separates out the key frame request from the frame ACK, and
moves it into the memory allocation logic. Following a memory allocation
or re-allocation, a key frame will always need to be sent.

(cherry picked from commit b728306f201b6e4a275c0760d3e93b0a2bd05885)